Fixed int/unsigned int issue

+++ algorithms/profile_model/gaussian_rs/mask_calculator.h
@@ -201,7 +201,7 @@ namespace gaussian_rs {
           double dxy4 = (gxy4[0]*gxy4[0] + gxy4[1]*gxy4[1]) * delta_b_r2;
           double dxy = std::min(std::min(dxy1, dxy2), std::min(dxy3, dxy4));
           for (std::size_t k = 0; k < zsize; ++k) {
-            if (z0 + k >= index0_ && z0 + k < index1_) {
+            if (z0 + (int)k >= index0_ && z0 + (int)k < index1_) {
               /* double gz1 = cs.from_rotation_angle_fast(phi0_ + (z0 + k - index0_) * dphi_); */
               /* double gz2 = cs.from_rotation_angle_fast(phi0_ + (z0 + k + 1 - index0_) * dphi_); */
               /* double gz = std::abs(gz1) < std::abs(gz2) ? gz1 : gz2; */
